Refactored to //model
