+++ ui/src/components/admin/index.js
@@ -4,7 +4,7 @@ import { Route } from 'react-router-dom';
 import AdminProvider from 'providers/AdminProvider';
 import AdminNav from './AdminNav';
 import ModelsManager from './ModelsManager';
-import { ModelSingle } from './ModelUpload';
+import { ModelSingle } from './Model';
 import UsersManage from './UsersManage';
 import { Col } from 'theme/system';
 import { AdminMain } from 'theme/adminStyles';
@@ -16,7 +16,7 @@ export default ({ location }) => (
       <AdminMain>
         <Route path="/admin/manage-users" render={() => <UsersManage />} />
         <Route path="/admin/manage-models" render={() => <ModelsManager />} />
-        <Route path="/admin/model-upload-single" render={() => <ModelSingle />} />
+        <Route path="/admin/model" render={() => <ModelSingle />} />
       </AdminMain>
     </Col>
   </AdminProvider>
